FIX: avoid size_t in variables set by command line to avoid across-platform type confusion
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@171 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/apps/mrbias.cxx
@@ -57,7 +57,7 @@ float ThresholdForegroundMax = FLT_MAX;
 bool ThresholdForegroundFlag = false;
 const char* FNameMaskImage = NULL;
 float SamplingDensity = 0;
-size_t NumberOfHistogramBins = 256;
+unsigned int NumberOfHistogramBins = 256;
 
 const char* FNameInputImage = NULL;
 const char* FNameOutputImage = NULL;
